Se añadio la migracion de areas y se modifico proyectos.

+++ app/Http/Controllers/ProjectController.php
@@ -1,8 +1,11 @@
 <?php
 
 namespace App\Http\Controllers;
+
+use App\Models\Area;
 use App\Models\Project;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Validation\ValidationException;
 
 class ProjectController extends Controller
@@ -15,7 +18,7 @@ class ProjectController extends Controller
         return inertia(
             'Project/IndexProject',
             [
-                'projects' => Project::all()
+                'projects' => Project::with('area')->get(),
             ]);
     }
 
@@ -24,7 +27,10 @@ class ProjectController extends Controller
      */
     public function create()
     {
-        return inertia('Project/CreateProject');
+        abort_if(!Auth::user()->can('create projects'), 403, 'No tienes los permisos necesarios para ver esta pagina.');
+        return inertia('Project/CreateProject', [
+            'areas' => Area::all(),
+        ]);
     }
 
     /**
@@ -41,7 +47,11 @@ class ProjectController extends Controller
                     'start_date' => 'required',
                     'end_date' => 'required',
                     'status' => 'required|string',
+                    'area_id' => 'required|integer|exists:areas,id',
         ]);
+        $company = $request->user()->companies()->first();
+        $validated['company_id'] = $company->id ?? null;
+
         $request->user()->projects()->create($validated);
 
             //los datos se guardan en BD
@@ -51,6 +61,7 @@ class ProjectController extends Controller
         }catch(\Exception $e){
             //si los datos no se guardan, se muestra mensaje de error
             return redirect()->back()->with('error', 'Error al generar proyecto.')->withInput();//withInput mantiene los datos del form
+
         }
     }
 
@@ -59,10 +70,11 @@ class ProjectController extends Controller
      */
     public function show(Project $project)
     {
+        $project->load('area');
         return inertia(
             'Project/ShowProject',
             [
-                'project' => $project
+                'project' => $project,
             ]);
     }
 
@@ -71,10 +83,13 @@ class ProjectController extends Controller
      */
     public function edit(Project $project)
     {
+
+        abort_if(!Auth::user()->can('edit projects'), 403, 'No tienes los permisos necesarios para realizar esta acción.');
         return inertia(
             'Project/EditProject',
             [
-                'project' => $project
+                'project' => $project,
+                'areas' => Area::all(),
             ]);
     }
 
@@ -92,6 +107,7 @@ class ProjectController extends Controller
                     'start_date' => 'required',
                     'end_date' => 'required',
                     'status' => 'required|string',
+                    'area_id' => 'required|integer|exists:areas,id',
         ]);
         $project->update($validated);
 
@@ -110,6 +126,7 @@ class ProjectController extends Controller
      */
     public function destroy(Project $project)
     {
+        abort_if(!Auth::user()->can('delete projects'), 403, 'No tienes los permisos necesarios para realizar esta acción.');
         $project->delete();
         return redirect()->route('projects.index')->with('success', 'Proyecto eliminado.');
     }
